Moved get_compiler_executable method to SolidityCompiler trait
