+++ crates/compiler/Cargo.toml
@@ -11,6 +11,8 @@ rust-version.workspace = true
 [dependencies]
 anyhow = { workspace = true }
 revive-solc-json-interface = { workspace = true }
+revive-dt-config = { workspace = true }
+revive-dt-solc-binaries = { workspace = true }
 revive-common = { workspace = true }
 semver = { workspace = true }
 serde_json = { workspace = true }
