+++ crates/compiler/src/solc.rs
@@ -7,6 +7,8 @@ use std::{
 };
 
 use crate::{CompilerInput, CompilerOutput, SolidityCompiler};
+use revive_dt_config::Arguments;
+use revive_dt_solc_binaries::download_solc;
 
 pub struct Solc {
     solc_path: PathBuf,
@@ -38,5 +40,10 @@ impl SolidityCompiler for Solc {
 
     fn new(solc_path: PathBuf) -> Self {
         Self { solc_path }
+    }
+    
+    fn get_compiler_executable(config: &Arguments, version: semver::Version) -> anyhow::Result<PathBuf> {
+        let path = download_solc(config.directory(), version, config.wasm)?;
+        Ok(path)
     }
 }
