+++ crates/core/src/driver/mod.rs
@@ -49,7 +49,8 @@ where
             compiler = compiler.with_source(file)?;
         }
 
-        let compiler_path = T::get_compiler_executable(self.config, version)?;
+        let compiler_path = T::Compiler::get_compiler_executable(self.config, version)?;
+
 
         let output = compiler
             .solc_optimizer(mode.solc_optimize())
