+++ crates/core/src/lib.rs
@@ -4,13 +4,8 @@
 //! provides a helper utilty to execute tests.
 
 use revive_dt_compiler::{SolidityCompiler, solc, revive_resolc};
-use revive_dt_config::Arguments;
 use revive_dt_node::geth;
 use revive_dt_node_interaction::EthereumNode;
-use revive_dt_solc_binaries::download_solc;
-
-use std::path::PathBuf;
-use semver::Version;
 
 pub mod driver;
 
@@ -21,7 +16,6 @@ pub trait Platform {
     type Blockchain: EthereumNode;
     type Compiler: SolidityCompiler;
 
-    fn get_compiler_executable(config: &Arguments, version: Version) -> anyhow::Result<PathBuf>;
 }
 
 #[derive(Default)]
@@ -30,12 +24,6 @@ pub struct Geth;
 impl Platform for Geth {
     type Blockchain = geth::Instance;
     type Compiler = solc::Solc;
-    
-    fn get_compiler_executable(config: &Arguments, version: Version) -> anyhow::Result<PathBuf> {
-
-        let path = download_solc(config.directory(), version, config.wasm)?;
-        Ok(path)
-    }
 }
 
 #[derive(Default)]
@@ -44,13 +32,5 @@ pub struct Kitchensink;
 impl Platform for Kitchensink {
     type Blockchain = geth::Instance;
     type Compiler = revive_resolc::Resolc;
-    
-    fn get_compiler_executable(config: &Arguments, _version: Version) -> anyhow::Result<PathBuf> {
-
-        if !config.resolc.as_os_str().is_empty() {
-            return Ok(config.resolc.clone());
-        }
-
-        Ok(PathBuf::from("resolc"))
-    }
+   
 }
